+++ CHANGELOG.md
@@ -20,10 +20,13 @@ All notable changes to the "cosmy-wasmy" extension will be documented in this fi
 
 ## [Unreleased]
 
+### Added
+
+- Implemented [#3](https://github.com/spoo-bar/cosmy-wasmy/issues/3) & [#4](https://github.com/spoo-bar/cosmy-wasmy/issues/4) - The smart contracts in the Contract view can now be sorted alphabetically, or with CodeID, or in imported order
+
 ### Changed
 
 - Implemented [#2](https://github.com/spoo-bar/cosmy-wasmy/issues/2) - Show faucet icon for account only when faucet has been set up for that chain in the settings
-
 
 
 ## [v0.2.0] - 15 June 2022